fn-validate-insert-arquetype/requirements.txt
functions-framework==3.8.2
google-cloud-storage==3.0.0
google-cloud-bigquery==3.30.0
